requirements.txt
PyYaml
six
py-timeparser
mock
pytest